Time: 34 ms (7.03%) | Memory: 47.8 MB (74.36%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.java
@@ -2,7 +2,7 @@ class Solution {
     public int find(int[] piles,int k){
         int total=0;
         for(int i=0;i<piles.length;i++){
-            total+=Math.ceil((double)piles[i]/(double)k);
+            total+=Math.ceil((double)piles[i]/k);
         }
         return total;
     }
